Refactored Texts to use TextTheme
