+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
 import 'app_module.dart';
-import 'app_widget.dart';
+import 'presenter/app_widget.dart';
 
 void main() {
   runApp(ModularApp(module: AppModule(), child: const AppWidget()));
